Refactor WebsocketComponent
